feat: enable REST query API by default
The REST controllers don't distinguish between regular and query API anymore.
All V2 C8 REST endpoints are enabled by default and can only be disabled together.

The configuration option "camunda.rest.query.enabled" will not be considered anymore.
