+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/ProcessInstanceController.java
@@ -7,6 +7,9 @@
  */
 package io.camunda.zeebe.gateway.rest.controller;
 
+import static io.camunda.zeebe.gateway.rest.RestErrorMapper.mapErrorToResponse;
+
+import io.camunda.search.query.ProcessInstanceQuery;
 import io.camunda.security.configuration.MultiTenancyConfiguration;
 import io.camunda.service.ProcessInstanceServices;
 import io.camunda.service.ProcessInstanceServices.ProcessInstanceCancelRequest;
@@ -17,12 +20,17 @@ import io.camunda.zeebe.gateway.protocol.rest.CancelProcessInstanceRequest;
 import io.camunda.zeebe.gateway.protocol.rest.CreateProcessInstanceRequest;
 import io.camunda.zeebe.gateway.protocol.rest.MigrateProcessInstanceRequest;
 import io.camunda.zeebe.gateway.protocol.rest.ModifyProcessInstanceRequest;
+import io.camunda.zeebe.gateway.protocol.rest.ProcessInstanceSearchQueryRequest;
+import io.camunda.zeebe.gateway.protocol.rest.ProcessInstanceSearchQueryResponse;
 import io.camunda.zeebe.gateway.rest.RequestMapper;
 import io.camunda.zeebe.gateway.rest.ResponseMapper;
 import io.camunda.zeebe.gateway.rest.RestErrorMapper;
+import io.camunda.zeebe.gateway.rest.SearchQueryRequestMapper;
+import io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper;
 import java.util.concurrent.CompletableFuture;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -82,6 +90,48 @@ public class ProcessInstanceController {
       @RequestBody final ModifyProcessInstanceRequest modifyRequest) {
     return RequestMapper.toModifyProcessInstance(processInstanceKey, modifyRequest)
         .fold(RestErrorMapper::mapProblemToCompletedResponse, this::modifyProcessInstance);
+  }
+
+  @PostMapping(
+      path = "/search",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE},
+      consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<ProcessInstanceSearchQueryResponse> searchProcessInstances(
+      @RequestBody(required = false) final ProcessInstanceSearchQueryRequest query) {
+    return SearchQueryRequestMapper.toProcessInstanceQuery(query)
+        .fold(RestErrorMapper::mapProblemToResponse, this::search);
+  }
+
+  @GetMapping(
+      path = "/{processInstanceKey}",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE})
+  public ResponseEntity<Object> getByKey(
+      @PathVariable("processInstanceKey") final Long processInstanceKey) {
+    try {
+      // Success case: Return the left side with the ProcessInstanceItem wrapped in ResponseEntity
+      return ResponseEntity.ok()
+          .body(
+              SearchQueryResponseMapper.toProcessInstance(
+                  processInstanceServices
+                      .withAuthentication(RequestMapper.getAuthentication())
+                      .getByKey(processInstanceKey)));
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
+  }
+
+  private ResponseEntity<ProcessInstanceSearchQueryResponse> search(
+      final ProcessInstanceQuery query) {
+    try {
+      final var result =
+          processInstanceServices
+              .withAuthentication(RequestMapper.getAuthentication())
+              .search(query);
+      return ResponseEntity.ok(
+          SearchQueryResponseMapper.toProcessInstanceSearchQueryResponse(result));
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
   }
 
   private CompletableFuture<ResponseEntity<Object>> createProcessInstance(
