+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/UserTaskController.java
@@ -7,19 +7,35 @@
  */
 package io.camunda.zeebe.gateway.rest.controller;
 
+import static io.camunda.zeebe.gateway.rest.RestErrorMapper.mapErrorToResponse;
+
+import io.camunda.search.entities.FormEntity;
+import io.camunda.search.query.UserTaskQuery;
+import io.camunda.search.query.VariableQuery;
 import io.camunda.service.UserTaskServices;
+import io.camunda.zeebe.gateway.protocol.rest.FormItem;
 import io.camunda.zeebe.gateway.protocol.rest.UserTaskAssignmentRequest;
 import io.camunda.zeebe.gateway.protocol.rest.UserTaskCompletionRequest;
+import io.camunda.zeebe.gateway.protocol.rest.UserTaskItem;
+import io.camunda.zeebe.gateway.protocol.rest.UserTaskSearchQueryRequest;
+import io.camunda.zeebe.gateway.protocol.rest.UserTaskSearchQueryResponse;
 import io.camunda.zeebe.gateway.protocol.rest.UserTaskUpdateRequest;
+import io.camunda.zeebe.gateway.protocol.rest.UserTaskVariableSearchQueryRequest;
+import io.camunda.zeebe.gateway.protocol.rest.VariableSearchQueryResponse;
 import io.camunda.zeebe.gateway.rest.RequestMapper;
 import io.camunda.zeebe.gateway.rest.RequestMapper.AssignUserTaskRequest;
 import io.camunda.zeebe.gateway.rest.RequestMapper.CompleteUserTaskRequest;
 import io.camunda.zeebe.gateway.rest.RequestMapper.UpdateUserTaskRequest;
 import io.camunda.zeebe.gateway.rest.RestErrorMapper;
+import io.camunda.zeebe.gateway.rest.SearchQueryRequestMapper;
+import io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper;
+import io.camunda.zeebe.gateway.rest.cache.ProcessCache;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -31,9 +47,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class UserTaskController {
 
   private final UserTaskServices userTaskServices;
+  private final ProcessCache processCache;
 
-  public UserTaskController(final UserTaskServices userTaskServices) {
+  public UserTaskController(
+      final UserTaskServices userTaskServices, final ProcessCache processCache) {
     this.userTaskServices = userTaskServices;
+    this.processCache = processCache;
   }
 
   @PostMapping(
@@ -77,6 +96,91 @@ public class UserTaskController {
 
     return RequestMapper.toUserTaskUpdateRequest(updateRequest, userTaskKey)
         .fold(RestErrorMapper::mapProblemToCompletedResponse, this::updateUserTask);
+  }
+
+  @PostMapping(
+      path = "/search",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE},
+      consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<UserTaskSearchQueryResponse> searchUserTasks(
+      @RequestBody(required = false) final UserTaskSearchQueryRequest query) {
+    return SearchQueryRequestMapper.toUserTaskQuery(query)
+        .fold(RestErrorMapper::mapProblemToResponse, this::search);
+  }
+
+  @GetMapping(
+      path = "/{userTaskKey}",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE})
+  public ResponseEntity<UserTaskItem> getByKey(
+      @PathVariable("userTaskKey") final Long userTaskKey) {
+    try {
+      final var userTask =
+          userTaskServices
+              .withAuthentication(RequestMapper.getAuthentication())
+              .getByKey(userTaskKey);
+      final var name = processCache.getUserTaskName(userTask);
+      return ResponseEntity.ok().body(SearchQueryResponseMapper.toUserTask(userTask, name));
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
+  }
+
+  @GetMapping(
+      path = "/{userTaskKey}/form",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE})
+  public ResponseEntity<FormItem> getFormByUserTaskKey(
+      @PathVariable("userTaskKey") final long userTaskKey) {
+    try {
+      final Optional<FormEntity> form =
+          userTaskServices
+              .withAuthentication(RequestMapper.getAuthentication())
+              .getUserTaskForm(userTaskKey);
+      return form.map(SearchQueryResponseMapper::toFormItem)
+          .map(ResponseEntity::ok)
+          .orElseGet(() -> ResponseEntity.noContent().build());
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
+  }
+
+  @PostMapping(
+      path = "/{userTaskKey}/variables/search",
+      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE},
+      consumes = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<VariableSearchQueryResponse> searchVariables(
+      @PathVariable("userTaskKey") final long userTaskKey,
+      @RequestBody(required = false)
+          final UserTaskVariableSearchQueryRequest userTaskVariablesSearchQueryRequest) {
+
+    return SearchQueryRequestMapper.toUserTaskVariableQuery(userTaskVariablesSearchQueryRequest)
+        .fold(
+            RestErrorMapper::mapProblemToResponse,
+            query -> searchUserTaskVariableQuery(userTaskKey, query));
+  }
+
+  private ResponseEntity<UserTaskSearchQueryResponse> search(final UserTaskQuery query) {
+    try {
+      final var result =
+          userTaskServices.withAuthentication(RequestMapper.getAuthentication()).search(query);
+      final var processCacheItems = processCache.getUserTaskNames(result.items());
+      return ResponseEntity.ok(
+          SearchQueryResponseMapper.toUserTaskSearchQueryResponse(result, processCacheItems));
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
+  }
+
+  private ResponseEntity<VariableSearchQueryResponse> searchUserTaskVariableQuery(
+      final long userTaskKey, final VariableQuery query) {
+    try {
+      final var result =
+          userTaskServices
+              .withAuthentication(RequestMapper.getAuthentication())
+              .searchUserTaskVariables(userTaskKey, query);
+      return ResponseEntity.ok(SearchQueryResponseMapper.toVariableSearchQueryResponse(result));
+    } catch (final Exception e) {
+      return mapErrorToResponse(e);
+    }
   }
 
   private CompletableFuture<ResponseEntity<Object>> assignUserTask(
