--- zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/usermanagement/RoleQueryController.java
@@ -1,67 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.zeebe.gateway.rest.controller.usermanagement;
-
-import io.camunda.search.query.RoleQuery;
-import io.camunda.service.RoleServices;
-import io.camunda.zeebe.gateway.protocol.rest.RoleSearchQueryRequest;
-import io.camunda.zeebe.gateway.protocol.rest.RoleSearchQueryResponse;
-import io.camunda.zeebe.gateway.rest.RequestMapper;
-import io.camunda.zeebe.gateway.rest.RestErrorMapper;
-import io.camunda.zeebe.gateway.rest.SearchQueryRequestMapper;
-import io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper;
-import io.camunda.zeebe.gateway.rest.controller.CamundaRestQueryController;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@CamundaRestQueryController
-@RequestMapping("/v2/roles")
-public class RoleQueryController {
-  private final RoleServices roleServices;
-
-  public RoleQueryController(final RoleServices roleServices) {
-    this.roleServices = roleServices;
-  }
-
-  @GetMapping(
-      path = "/{roleKey}",
-      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE})
-  public ResponseEntity<Object> getRole(@PathVariable final long roleKey) {
-    try {
-      return ResponseEntity.ok()
-          .body(SearchQueryResponseMapper.toRole(roleServices.getRole(roleKey)));
-    } catch (final Exception exception) {
-      return RestErrorMapper.mapErrorToResponse(exception);
-    }
-  }
-
-  @PostMapping(
-      path = "/search",
-      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE},
-      consumes = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<RoleSearchQueryResponse> searchRoles(
-      @RequestBody(required = false) final RoleSearchQueryRequest query) {
-    return SearchQueryRequestMapper.toRoleQuery(query)
-        .fold(RestErrorMapper::mapProblemToResponse, this::search);
-  }
-
-  private ResponseEntity<RoleSearchQueryResponse> search(final RoleQuery query) {
-    try {
-      final var result =
-          roleServices.withAuthentication(RequestMapper.getAuthentication()).search(query);
-      return ResponseEntity.ok(SearchQueryResponseMapper.toRoleSearchQueryResponse(result));
-    } catch (final Exception e) {
-      return RestErrorMapper.mapErrorToResponse(e);
-    }
-  }
-}
