+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/UserTaskQueryControllerTest.java
@@ -50,7 +50,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 
-@WebMvcTest(value = UserTaskQueryController.class, properties = "camunda.rest.query.enabled=true")
+@WebMvcTest(value = UserTaskController.class)
 @Import(JacksonConfig.class)
 public class UserTaskQueryControllerTest extends RestControllerTest {
